finished

+++ Models/BotModels.cs
@@ -10,17 +10,17 @@ namespace EchoBot.Models
 
         public struct MainObject
         {
-            public double temp;
-            public double feels_like;
-            public double temp_min;
-            public double temp_max;
+            public float temp;
+            public float feels_like;
+            public float temp_min;
+            public float temp_max;
             public int pressure;
             public int humidity;
         }
 
         public struct WindObject
         {
-            public double speed;
+            public float speed;
         }
 
         public struct SysObject
